dead letter queue, rate limiting, concurrency groups, tracing, bug fixes, clippy clean
Dead letter queue: failed jobs captured with error details, replayable.
Rate limiting: per-job sliding window (max_executions/window_seconds).
Concurrency groups: named groups with shared max_concurrent limit.
Structured tracing: spans on scheduler loop and job execution, optional
OpenTelemetry OTLP export behind 'otel' feature flag.

Fixed APScheduler bugs:
- Stale lease cleanup on scheduler startup (new cleanup_stale_leases trait method)
- Jobs added during shutdown now rejected with SchedulerShuttingDown error
- Verified: multi-scheduler race condition prevented by FOR UPDATE SKIP LOCKED
- Verified: DST fall-back handled correctly (no infinite loop)
- Verified: all internal datetimes are UTC (no naive/aware confusion)

+++ crates/triggers/src/calendar.rs
@@ -78,8 +78,8 @@ impl CalendarIntervalTrigger {
     /// Add the calendar interval to a date, clamping the day to the end of
     /// the target month if necessary.
     fn add_interval(&self, date: NaiveDate) -> Option<NaiveDate> {
-        let total_months = self.years as i32 * 12 + self.months as i32;
-        let total_days = self.weeks as i32 * 7 + self.days as i32;
+        let total_months = self.years * 12 + self.months;
+        let total_days = self.weeks * 7 + self.days;
 
         let mut new_date = date;
